+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
  - Add ability for Actions to work on Function Controllers
  - Add LCD Backlight On/Off Actions to LCD Functions ([#1089](https://github.com/kizniche/mycodo/issues/1089))
+ - Add Input: SHT2x (using sht20 library with changeable temperature resolution)
 
 ### Miscellaneous